pyproject.toml
[project]
name = "visual-rag"
version = "0.1.0"
description = "Visual RAG with ColPali and Milvus"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pymilvus>=2.6.4",
    "pymupdf>=1.23.0",
    "pillow>=10.2.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "loguru>=0.7.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.6",
    "colpali-engine>=0.3.13",
    "accelerate>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (real API calls)",
    "unit: marks tests as unit tests (mocked)"
]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

